rm bjs stuff for now
